trash/inner_docker/requirements.txt
playwright==1.42.0
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio
